Use a better "underline" character than underscore
Works around the hacky underlining which just renders an underscore
character behind the letter. This underscore character looks terrible on
many fonts, (e.g. Iosevka) and is actually confusing since it makes
spaces seem like underscores.

I just replaced that character with a better underline character that's
present in the Iosevka Nerd Font. This actually looks exactly like an
underscore would look (except for the fact that it doesn't break the
line for letters like y. If the user doesn't have a font that includes
this character (it's in Unicode private usage plane), then underline
won't be displayed. This is still better than layering an underscore.